Seat Allotment System
This project Entrance Seat Allotment System is windows application in
which students can register with their rank number for the entrance
examination and the administrator can allot the seats for the students.
Administrator can add the college details and he batch details. Using
this software the entrance seat allotment became easier and can be
implemented using system. The main advantage of the project is the
computerization of the entrance seat allotment process. Administrator
has the power for the allotment. He can add the allotted seats into a
file and the details are saved into the system. The total time for the
entrance allotment became lesser and the allotment process became
faster.